refactor: move rotateNotificationLogs to its own file
