+++ src/game-logic/reducers/rotateNotificationLogs.js
@@ -1,0 +1,35 @@
+import { NOTIFICATION_LOG_SIZE } from '../../constants'
+
+/**
+ * @param {farmhand.state} state
+ * @returns {farmhand.state}
+ */
+export const rotateNotificationLogs = state => {
+  const notificationLog = [...state.notificationLog]
+
+  const { dayCount, newDayNotifications } = state
+
+  const notifications = {
+    error: [],
+    info: [],
+    success: [],
+    warning: [],
+  }
+
+  newDayNotifications.forEach(({ message, severity }) =>
+    notifications[severity].push(message)
+  )
+
+  newDayNotifications.length &&
+    notificationLog.unshift({
+      day: dayCount,
+      notifications,
+    })
+
+  notificationLog.length = Math.min(
+    notificationLog.length,
+    NOTIFICATION_LOG_SIZE
+  )
+
+  return { ...state, notificationLog }
+}
